+++ src/adam/model/__init__.py
@@ -1,6 +1,6 @@
-from .factories.abc_factories import Joint, Link, ModelFactory
-from .factories.std_joint import StdJoint
-from .factories.std_link import StdLink
-from .factories.std_model import URDFModelFactory
+from .abc_factories import Joint, Link, ModelFactory
 from .model import Model
+from .std_factories.std_joint import StdJoint
+from .std_factories.std_link import StdLink
+from .std_factories.std_model import URDFModelFactory
 from .tree import Node, Tree
