+++ src/adam/model/abc_factories.py
